2.8.10 Extended database support, improved LpCode

See merge request intellisrc/common!16

+++ modules/db/src/test/groovy/com/intellisrc/db/jdbc/JDBCTest.groovy
@@ -24,7 +24,7 @@ abstract class JDBCTest extends Specification {
     boolean shouldSkip() {
         JDBC jdbc = this.getDB()
         boolean skip = jdbc instanceof JDBCServer
-            && (ci || ((jdbc as JDBCServer).port &&! LocalHost.hasOpenPort((jdbc as JDBCServer).port)))
+            && (ci || ((jdbc as JDBCServer).hostname &&! LocalHost.hasOpenPort((jdbc as JDBCServer).port)))
         if(skip) {
             Log.w("Test skipped for : %s (environment not ready)", jdbc.class.simpleName)
         }
@@ -56,8 +56,12 @@ abstract class JDBCTest extends Specification {
 
     def setup() {
         DB db = getDB().connect()
-        db.dropAllTables()
-        db.clearCache()
+        if(db.openIfClosed()) {
+            db.dropAllTables()
+            db.clearCache()
+        } else {
+            Log.w("Connection failed")
+        }
         db.close()
     }
 
@@ -84,17 +88,17 @@ abstract class JDBCTest extends Specification {
         then:
             assert getTableCreate(table) || getTableCreateMulti(table) : "Missing table create"
         then: "Create table"
-            assert getTableCreate(table) ? db.setSQL(getTableCreate(table)) : db.setSQL(getTableCreateMulti(table))
+            assert getTableCreate(table) ? db.setSQL(getTableCreate(table)) : db.setSQL(getTableCreateMulti(table)) : "Unable to create table"
         then: "List tables"
-            assert ! db.tables.empty
+            assert ! db.tables.empty : "Tables not found"
         then: "Must be open"
-            assert db.opened
-            assert ! db.closed
+            assert db.opened : "Database not opened"
+            assert ! db.closed : "Database was closed"
         when: "Table info"
             List<ColumnInfo> info = db.table(table).info()
             new TableMaker(info.collect { it.toMap() }).print()
         then: "Must have information"
-            assert ! info.empty
+            assert ! info.empty : "No information found"
         then: "Must have a primary key"
             assert info.find { it.primaryKey }
         then: "Table must exists"
@@ -103,11 +107,15 @@ abstract class JDBCTest extends Specification {
             assert db.table(table).insert(
                 [ name : "Ubuntu",     active: true,    updated: setDate("2022-08-01"),     version: 3.7 ],
             )
+        when:
+            new TableMaker(db.table(table).get().toListMap()).print()
             assert db.lastID == 1
         then: "Insert Second"
             assert db.table(table).insert(
                 [ name : "Debian",     active: true,    updated: setDate("2022-09-01"),     version: 9.1 ],
             )
+        when:
+            new TableMaker(db.table(table).get().toListMap()).print()
             assert db.lastID == 2
         then: "Count all"
             assert db.table(table).count().get().toInt() == 2
